Don't change theme in external plugins
Introduced a new class for external plugins, so they can be skipped in yin_yang.py via a isinstance() check
